- trivial assignments
(does not compile, this commit is only supposed to reduce the noise on the next one.

+++ source/games/sw/src/game.h
@@ -969,8 +969,8 @@ struct USER
 
     double oz; // serialized copy of sprite.oz
     double z_tgt;
-    int16_t ceiling_dist;
-    int16_t floor_dist;
+    double ceiling_dist;
+    double floor_dist;
 
     int Flags;
     int Flags2;
